+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/PlaceObject3Tag.java
@@ -443,6 +443,68 @@ public class PlaceObject3Tag extends PlaceObjectTypeTag implements ASMSourceCont
     }
 
     @Override
+    public void getDataNoScript(SWFOutputStream sos) throws IOException {
+        sos.writeUB(1, 0); //no clip actions
+        sos.writeUB(1, placeFlagHasClipDepth ? 1 : 0);
+        sos.writeUB(1, placeFlagHasName ? 1 : 0);
+        sos.writeUB(1, placeFlagHasRatio ? 1 : 0);
+        sos.writeUB(1, placeFlagHasColorTransform ? 1 : 0);
+        sos.writeUB(1, placeFlagHasMatrix ? 1 : 0);
+        sos.writeUB(1, placeFlagHasCharacter ? 1 : 0);
+        sos.writeUB(1, placeFlagMove ? 1 : 0);
+        sos.writeUB(1, reserved ? 1 : 0);
+        sos.writeUB(1, placeFlagOpaqueBackground ? 1 : 0); //SWF11
+        sos.writeUB(1, placeFlagHasVisible ? 1 : 0); //SWF11
+        sos.writeUB(1, placeFlagHasImage ? 1 : 0);
+        sos.writeUB(1, placeFlagHasClassName ? 1 : 0);
+        sos.writeUB(1, placeFlagHasCacheAsBitmap ? 1 : 0);
+        sos.writeUB(1, placeFlagHasBlendMode ? 1 : 0);
+        sos.writeUB(1, placeFlagHasFilterList ? 1 : 0);
+        sos.writeUI16(depth);
+
+        if (placeFlagHasClassName) {
+            sos.writeString(className);
+        }
+        if (placeFlagHasCharacter) {
+            sos.writeUI16(characterId);
+        }
+        if (placeFlagHasMatrix) {
+            sos.writeMatrix(matrix);
+        }
+        if (placeFlagHasColorTransform) {
+            sos.writeCXFORMWITHALPHA(colorTransform);
+        }
+        if (placeFlagHasRatio) {
+            sos.writeUI16(ratio);
+        }
+        if (placeFlagHasName) {
+            sos.writeString(name);
+        }
+        if (placeFlagHasClipDepth) {
+            sos.writeUI16(clipDepth);
+        }
+        if (placeFlagHasFilterList) {
+            sos.writeFILTERLIST(surfaceFilterList);
+        }
+        if (placeFlagHasBlendMode) {
+            sos.writeUI8(blendMode);
+        }
+        if (placeFlagHasCacheAsBitmap) {
+            if (!bitmapCacheBug) {
+                sos.writeUI8(bitmapCache);
+            }
+        }
+        if (placeFlagHasVisible) {
+            sos.writeUI8(visible);
+        }
+        if (placeFlagOpaqueBackground) {
+            sos.writeRGBA(backgroundColor);
+        }
+    }
+    
+    
+
+    @Override
     public int getPlaceObjectNum() {
         return 3;
     }