+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/Tag.java
@@ -561,6 +561,18 @@ public abstract class Tag implements NeedsCharacters, Exportable, Serializable {
             sos.write(originalRange.getArray(), originalRange.getPos(), originalRange.getLength());
         }
     }
+    
+    /**
+     * Writes Tag value to the stream, ignoring all scripts
+     * @param sos
+     * @throws IOException 
+     */
+    public void writeTagNoScripts(SWFOutputStream sos) throws IOException {
+        byte[] newData = getDataNoScript();
+        byte[] newHeaderData = getHeader(newData.length);
+        sos.write(newHeaderData);
+        sos.write(newData);
+    }        
 
     /**
      * Clones the tag.
@@ -625,7 +637,7 @@ public abstract class Tag implements NeedsCharacters, Exportable, Serializable {
      * @throws IOException On I/O error
      */
     public abstract void getData(SWFOutputStream sos) throws IOException;
-
+      
     /**
      * Gets data bytes
      *
@@ -643,6 +655,30 @@ public abstract class Tag implements NeedsCharacters, Exportable, Serializable {
 
         try (SWFOutputStream sos = new SWFOutputStream(os, getVersion(), getCharset())) {
             getData(sos);
+            if (remainingData != null) {
+                sos.write(remainingData);
+            }
+        } catch (IOException e) {
+            throw new Error("This should never happen.", e);
+        }
+
+        return baos.toByteArray();
+    }
+    
+    /**
+     * Gets data bytes ignoring all scripts
+     * @param sos SWF output stream
+     * @throws IOException On I/O error
+     */
+    public void getDataNoScript(SWFOutputStream sos) throws IOException {
+        getData(sos);
+    }
+    
+    public byte[] getDataNoScript() {
+        ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        OutputStream os = baos;
+        try (SWFOutputStream sos = new SWFOutputStream(os, getVersion(), getCharset())) {
+            getDataNoScript(sos);
             if (remainingData != null) {
                 sos.write(remainingData);
             }