+++ src/com/jpexs/decompiler/flash/gui/player/ZoomPanel.java
@@ -79,7 +79,7 @@ public class ZoomPanel extends JPanel implements MediaDisplayListener {
         rulerButton.addActionListener(this::rulerActionPerformed);
         rulerButton.setToolTipText(AppStrings.translate("button.ruler.hint"));
         rulerButton.setSelected(Configuration.showRuler.get());
-
+        
         Configuration.showRuler.addListener(new ConfigurationItemChangeListener<Boolean>() {
             @Override
             public void configurationItemChanged(Boolean newValue) {
@@ -114,7 +114,7 @@ public class ZoomPanel extends JPanel implements MediaDisplayListener {
         guidesOptionsButton.setToolTipText(AppStrings.translate("button.guides_options"));
 
         snapOptionsButton = new SnapOptionsButton();
-        
+
         PopupButton gridButton = new PopupButton(View.getIcon("grid16")) {
             @Override
             protected JPopupMenu getPopupMenu() {
@@ -123,15 +123,15 @@ public class ZoomPanel extends JPanel implements MediaDisplayListener {
                 showGridCheckBoxMenuItem.addActionListener(ZoomPanel.this::showGridCheckBoxMenuItemActionPerformed);
                 showGridCheckBoxMenuItem.setSelected(Configuration.showGrid.get());
                 menu.add(showGridCheckBoxMenuItem);
-                
+
                 JMenuItem editGridMenuItem = new JMenuItem(AppStrings.translate("grid_options.edit"));
                 editGridMenuItem.addActionListener(ZoomPanel.this::editGridMenuItemActionPerformed);
                 menu.add(editGridMenuItem);
-                
+
                 return menu;
-            }            
+            }
         };
-        gridButton.setToolTipText(AppStrings.translate("button.grid_options"));       
+        gridButton.setToolTipText(AppStrings.translate("button.grid_options"));
 
         
         percentLabel.setToolTipText(AppStrings.translate("zoom.hint"));
@@ -152,11 +152,11 @@ public class ZoomPanel extends JPanel implements MediaDisplayListener {
                     } catch (NumberFormatException nfe) {
                         //ignore
                     }
-                    
+
                 }
-            }            
+            }
         });
-        
+
         setLayout(new FlowLayout());
         add(percentLabel);
         add(zoomInButton);
@@ -170,12 +170,12 @@ public class ZoomPanel extends JPanel implements MediaDisplayListener {
 
         display.addEventListener(this);
     }
-    
+
     
     private void editGridMenuItemActionPerformed(ActionEvent evt) {
         new GridDialog(Main.getDefaultDialogsOwner()).setVisible(true);
     }
-    
+
     private void showGridCheckBoxMenuItemActionPerformed(ActionEvent evt) {
         JCheckBoxMenuItem source = (JCheckBoxMenuItem) evt.getSource();
         Configuration.showGrid.set(source.isSelected());
@@ -194,7 +194,7 @@ public class ZoomPanel extends JPanel implements MediaDisplayListener {
     private void guidesEditActionPerformed(ActionEvent evt) {
         new GuidesDialog(Main.getDefaultDialogsOwner(), display).setVisible(true);
     }
-    
+
     private void guidesClearActionPerformed(ActionEvent evt) {
         display.clearGuides();
     }
@@ -295,7 +295,7 @@ public class ZoomPanel extends JPanel implements MediaDisplayListener {
                 zoomToFit = zoom.fit;
                 realZoom = zoom.value;
                 updateZoomDisplay();
-            }
+            }                                    
         });
     }